initialize database through function
